// src/app/models/soundac-keys.ts
export class SoundacKeys {

    constructor() {
        this.active = '';
        this.activePubkey = '';
        this.basic = '';
        this.basicPubkey = '';
        this.memo = '';
        this.memoPubkey = '';
        this.owner = '';
        this.ownerPubkey = '';
    }

    active: string;
    activePubkey: string;
    basic: string;
    basicPubkey: string;
    memo: string;
    memoPubkey: string;
    owner: string;
    ownerPubkey: string;

}